* playing with histogram and threshold detection

+++ embedded/common/src/serial_arduino.cpp
@@ -17,7 +17,7 @@ serial::serial() {
   SOFT_UART.begin(9600);
 }
 
-size_t serial::read(uint8_t *buffer, size_t buffer_size, uint16_t timeout) {
+int16_t serial::read(uint8_t *buffer, size_t buffer_size, uint16_t timeout) {
   SOFT_UART.setTimeout(timeout);
   size_t ret=SOFT_UART.readBytes((char*)buffer,buffer_size);
   crc::CRC8(&rx_crc8,buffer,ret);
